requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
openai==1.3.3
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.25.1
